swapped name for chat and contacts
